The oldest encryption method
Encode the program written on the screen by encrypting the sentence until the dot (.) Is typed on the keyboard. As an encryption basis, one of the oldest encryption methods, “shifting each letter by a certain amount”, has been taken as basis. According to this; The amount of translation, which is entered from the keyboard, will be replaced by 3 letters from each letter when 3 is selected. In other words, all A's are denoted by D, and all K's are denoted by the letter N. When translated, the letters that pass the last letter are returned to the beginning and continue from there.